Add SwingWindow and SwingDialog functions that accept an `init` parameter.
